% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedIfElseToCaseWhen.R
\name{nestedIfElseToCaseWhen}
\alias{nestedIfElseToCaseWhen}
\title{Tidy ifelse statements}
\usage{
nestedIfElseToCaseWhen()
}
\description{
Convert nested \code{ifelse()} statements to \code{case_when()}.
}
